+++ src/java/com/javarush/tchaban/cryptoanalyser/FileProcessingException.java
@@ -1,6 +1,7 @@
 package com.javarush.tchaban.cryptoanalyser;
 
 public class FileProcessingException extends RuntimeException {
+
     String message;
 
     public FileProcessingException() {
