+++ src/java/com/javarush/tchaban/cryptoanalyser/FileProcessor.java
@@ -9,30 +9,24 @@ import java.util.List;
 public class FileProcessor {
 
     public List<String> readFile(String fileForReading) {
-        if (Validator.isValidPath(fileForReading)) {
-            List<String> stringList;
-            try {
-                stringList = Files.readAllLines(Path.of(fileForReading));
-            } catch (IOException ex) {
-                throw new FileProcessingException("Problem with file reading ", ex);
-            }
-            return stringList;
-        } else {
-            throw new IllegalArgumentException("Incorrect path to file " + fileForReading);
+        List<String> listAfterReading;
+        try {
+            Path path = Validator.returnValidPath(fileForReading);
+            listAfterReading = Files.readAllLines(path);
+        } catch (IOException ex) {
+            throw new FileProcessingException("Ошибка чтения файла " + fileForReading, ex);
         }
-    }
+        return listAfterReading;
+}
+
 
     public void writeToFile(List<String> stringList, String fileForWriting) {
-        if(Validator.isValidPath(fileForWriting)) {
             try {
-                Files.write(Path.of(fileForWriting), stringList);
+                Path path = Validator.returnValidPath(fileForWriting);
+                Files.write(path, stringList);
             } catch (IOException ex) {
-                throw new FileProcessingException("Problem with file writing ", ex);
+                throw new FileProcessingException("Ошибка записи файла " + fileForWriting, ex);
             }
-        } else {
-            throw new IllegalArgumentException("Incorrect path to file " + fileForWriting);
-        }
-
 
     }
